[Feature] Updated building data and UI components
- Removed unused `BuildingRegistryUpdater` class
- Updated `Building` model to use `Double` for height
- Adjusted `WikiBuildingPhoto` to scale and clip images
- Enhanced `BuildingDetailView` layout for better readability

+++ Stamped! A City Passport/Models/Building.swift
@@ -19,7 +19,7 @@ struct Building: Identifiable, Hashable, Codable {
     let newUse: String
     let buildingStyle: String
     let numberOfStories: Int
-    let height: Int
+    let height: Double
     let foodSpots: [String]
     let currency: String
     // Optional geographic coordinates — preferred if present in the registry JSON
